Fixed some bugs and finished the implementation for insertion in a quad tree.
The previous defined shape for a rectangle was bad in a way that we were
storing half the width and height using integers, which would only allow
for even sized rectangles. The implementation was changed to a top left,
width and height representation to allow odd sized rectangles. That was
a major bug, specially since we are working with geometrical data.
To reflect those changes, even the test files had to change, since they were
using the old representation for rectangles.

This commit also finishes the implementation for quad tree element insertion,
along with the inclusion of casserts to help us detect programming errors
inside the structure. We also corrected a bug caused by the split of a
rectangle that happened because of the impossibility to represent odd sized
rectangles in our previous implementation.