Setting targets to compile with -Wall -Werror and changing a little bit the interface for absolute difference.
Absolute difference was made into a customization point to allow users to
overwrite it for their classes. It now better specifies it's contracts and
uses SFINAE to correctly select the implementation. TODO: better specify
a string like type and implement string difference operation.
To reflect the changes, some adjustments had to be made in the test file,
specially regarding errors when comparing signed and unsigned values (which
were silently lurking in the code, waiting to crash).

Started adding compile features to targets on test directory to make sure
we can compile the code without any warnings and without invoking
UB.