Adding yet more warning flags and changing a little bit the interface of abs
Interface for abs changed. Now it always returns an unsigned value and in
the case the parameter passes being an unsigned value, it returns the
parameter itself. In case it receives an user defined type, as long
as it is possible to do a comparison to 0 and negate it, it behaves as
expected. Else, it is possible to customize it to your liking.

In order for targets to compile with warnings on, some parameters had to be
cast to void to trick the compiler and tests had to change because of
unsigned/signed comparisons.