Changing the interface for dimensional traits and getting to work on compact kd tree data structure.
The previous interface for dimension traits required the user to
do some metaprogramming himself, which made the definitions for user
defined type kd trees quite bad to work with. The changes still require
the user to supply information on the number of dimensions on a traits class
but other than that he just has to define get functions for his types.
TODO: see if there is a way to automatically detect the upmost index
for the user defined get function.

Compact KD Tree is a similar kd tree than the one we have implemented
in the project, with the difference it stores the elements in all nodes
and it is stored as an array of elements instead of pointers. But, in
contrast with our current implementation, which can support insertions
although it degrades the performance of the tree, it is a dense data
structure, meaning that adding elements would be a disaster.