Renaming some files to better show data structure meaning and implementing array kd tree copy operations.
Compact kd tree as a name didn't really convey that it was implemented
as an array, meaning it is a cache friendly structure. Meanwhile, the
default kd tree implementation needed a better explanation on the name,
since it didn't give a hint about the fact the points being stored
on the leaves. Also, I don't think it was a good idea to have that one
taking the name of the data structure, as it probably would mislead someone
into thinking that was the default implementation.

Implemented copy operations for the array kd tree. Now we just need to do
the same for the leaf kd tree.

Should work on documentation and class diagram for a while and then go
back to working on missing operations on current data structures.
Also thinking about adding a functional interface for the cases performance
doesn't matter that much.