Added some major improvements to nearest neighbor function API.
Nearest neighbor API required the user to do some metaprogramming on his
part from time to time, having to define templated functions to compare
types and objects. Now this is all hidden inside the classes. The user can
still use the API as before, but it is also possible to supply a type - type
function for each type, or handle specific dimensions differently even for the
same type defining type - type - dimension_t<Dimension> functions. All in
all, it is a lot more flexible and easier to use.

TODO: Reflect changes on doxygen.

Done some basic benchmarks for some structures. While the array kd tree
is building really fast, the node based kd tree is really slow to build
for some reason (comparing from 0.06s on array based kd trees to over
14s!!!!! on node based kd trees). While I do think building a dinamically
allocated array with only 1 call to malloc is definetelly better than building
a tree with lots of calls to the malloc allocator, I don't think this is the
main issue with the way we're building the node base kd tree and we can
certainly do better. That is something to keep track of in the future.